Moots: toggle buttons connect/disconnect outputs.
Proof of concept that the Moots module will work:
when a button is toggled on, polyphonic input is copied
to polyphonic output with the same number of channels.
When a button is toggled off, the output is set to
zero channels. This causes any cables connected to that
output to appears as "ghost" cables, and they have the
same effect as not being connected at all.