Moots: slewing option mostly finished.
The slewing option sounds like it is working,
but the menu checkboxes are not persisted with the saved patch yet.
I would also like to do more careful testing of the
exact waveform produced when fading in/out, but it
is hard to get the scope to graph this correctly when
it thinks its input signal cable keeps getting plugged/unplugged!