Moots slewing: eliminate special-case glitch.
There was an audio glitch in the following case:
A Moots controller was in manual push-button mode
(no gate cable connected), and the push-button was active.
Then use the context menu to turn ramping on.
This would cause a click because isGateActive[i] == false.

Now we always keep isGateActive[i] in sync with the activity
state of the controller, so when we toggle ramping, the
ramp always starts in the correct state and there is
no abrupt discontinuity in the output voltage level.