Fixed #2 - Moots: add slewing option for each control.
Finished implementing the slewing option.

Reworded the menu items to say "Anti-click ramping".
Hopefully this is clearer and easier to understand.

Documented the new menu options and added a screenshot.

Persist the menu options in the JSON blob for each module instance.
This way, after configuring a Moots module's slewing options,
they remain intact after reloading the patch or after duplicating
a module inside a patch.