Fixed CHANGELOG.md link. Updated CHANGELOG with actual release date.
In plugin.json, the link to CHANGELOG.md was incorrect,
but fortunately GitHub is smart enough to redirect it
from `master` to `main`. This is a low-priority fix that
can wait until the next release.

Version 2.0.1 went live on the VCV Rack Library today,
so I updated the release date accordingly in CHANGELOG.md.