Elastika: much more consistent output levels.
I was injecting audio by adding the input to
each input ball's velocity vector. This had the
effect of acting like an integrator, which caused
lower frequencies to be boosted and higher frequencies
to be attenuated. I replaced this with just setting
the input ball's velocity. Now the frequency response
is much more consistent, and does not vary greatly
with the stiffness setting like it did before.

There was also a level imbalance between the
left and right output channels. This was caused
by an asymmetry in how the stimulus and response
vectors were defined. Reworked so the mesh creator
defines its own stimulus/response vectors, not just the
input/output ball indices. It knows about its own
internal symmetry and can define the vectors accordingly.