Tone slider is now a direction angle slider.
Reworked the slider formerly used for TONE to
be a direction angle control. It affects both the
direction vector with which inputs are injected into
the mesh, and the direction vector with which audio
is extracted from the mesh.

No longer use an adjustable mixture of position/velocity
to produce output. Now, output is based on position only.
Overall, this gives a deeper sound, and it is my opinion
that users can adjust frequency response with external
filters if desired.