Disconnect LO/HI frequency filter knobs.
I'm moving away from having adjustable band-reject
filters in Elastika. This frees up the knobs at
the top of the panel.

I'm keeping a fixed lo-cut filter with a corner
frequency of 20 Hz. The LO/HI knobs still exist but they
don't do anything. I'm not sure what to do with the knobs,
or whether to remove them and do something else with
the top of the panel. Committing this just as a rewind
point for various potential experiments.