Elastika: add variable mass for a single ball.
I changed one knob at the top to adjust the mass of
a single ball, and the other knob to adjust
gravity strength. I'm not going to keep the gravity
adjustment because it's not much different from
adjusting a combination of stiffness and span.
There will be other changes in how mass works, both
in the simulation and in the UI... stay tuned!

Rework initialization code so `struct Elastika` holds
a `struct MeshAudioParameters` instead of lots of little
fields that get copied around. Eliminates a little bloat.