Elastika: slider for mass, knobs for tilt.
The knobs at the top of the display now independently
control the tilt angles of the input and output.
Before, both tilt angles were locked together and
controlled by a TILT slider.

What used to be the TILT slider is now the MASS slider.
It controls the "impurity masses" of two balls
over an exponential range 10^(-1) = 0.1 to 10^(+1) = 10.