Elastika: quiet the mesh after turning it off.
When the power button/gate turned off the mesh,
it was freezing the mesh in its current state.
All the balls remembered their positions and velocities.
Turning the mesh back on unfroze that state.
This didn't feel right. As a user, I expect the mesh
to go back to its original ground state when I thaw it.
That way it acts more like one would intuitively expect
a physical object to behave.

So now, as soon as the module completely turns off,
it quiets the mesh, meaning it puts all balls back
in their original positions and makes their velocities zero.