Elastika: allow full range for bipolar CV.
Based on the CV voltage standards:
https://vcvrack.com/manual/VoltageStandards#Levels

Control voltages (CV) can be bipolar [-5V, +5V] or
unipolar [0V, +10V]. I realized that a bipolar CV
was not capable of swinging the full range of my
sliders, even when the attenuverter was set to 100%.

I changed this so that this is now the case. It is critical
that I get this locked in before initial release,
so I won't break anybody's patches later.

So if a slider is all the way down, the attenuverter is 100%,
and a CV input is +5V, the result now is that the control
will act as if the slider is all the way up if there had
not be a CV applied.

It is still fine to use a unipolar CV, only it is likely
to get clamped when the attenuverter is outside the range
[-50%, +50%]. But that's just fine. I only want to make
sure I enable a bipolar CV to span the full range of control.

I had to fix two things:

- I was normalizing for 10V, not 5V as the supported amplitude.

- I forgot that some sliders have different ranges than others.
  Some are [0, +1], others are [-1, +1]. Now I adjust for
  the full range of each slider by including the factor
  (maxSlider - minSlider).

If someone complains later about how I did this, and they
can make a strong case why attenuverter settings aren't good
enough, I can always add right-click menu options to support
different CV ranges, while leaving [-5V, +5V] as the default range.