Fixed #7 - Sum polyphonic inputs to Elastika.
When a polyphonic cable is connected to any of Elastika's
inputs, whether the input is audio or CV, instead of using
only the first channel and ignoring the rest, add the
voltages across all channels and use the sum.

Elastika is not appropriate for true polyphony because
of its already-high CPU usage, but this is a good compromise
for processing polyphonic inputs.

It also allows for some interesting functionality where
polyphonic CV input jacks can act as unity mixers.