Moots: support bypass, sum polyphonic gate voltages.
Configure bypass to send each of the 5 input signals
to the corresponding output jack.
There is still an issue here, but this may be outside my control:
https://github.com/cosinekitty/sapphire/issues/9

Instead of using just the first channel's voltage
for gate inputs, sum the voltages on all the channels.
This allows for using each gate input as a unity mixer,
without changing the behavior on existing patches that
use a single-channel gate signal.