Elastika: self-correct NAN output every quarter second.
Elastika is not supposed to generate NAN output,
but there is a report that it sometimes does:

https://github.com/cosinekitty/sapphire/issues/8

As a final line of defense, check outputs about 4
times per second. If either is non-finite, quiet
the mesh, reset DC blockers, and send 0V to both
audio output channels.

The main intention is to prevent Elastika from
getting stuck in a NAN state forever.