Beginning refactor so I can use Elastika standalone.
I am starting to refactor Elastika so I can use it as a
standalone engine to generate sounds outside of VCV Rack.
This is the first baby step.
Moved class Slewer into a new header file that can be shared
by Moots and Elastika, without pulling anything in from
the VCV Rack SDK.