Tube Unit: added reflection decay, reflection angle knobs.
Deleted the realtime project for Tube Unit, because
maintaining it is just going to slow me down at this
point. It was helpful to get started while I was on
vacation, but I find that turning knobs in VCV Rack
is far better for experimentation than typing in numbers
on the command line.