Increased interpolator table size for better accuracy.
Unit tests were failing because the interpolation table
was too small. Increased its size, and improved floating
point accuracy by making the denominator a power of two.
Because we divide by (nsegments-1), it's better to make
nsegments 2**N + 1. In this case it is 0x8001.

Added a FIXFIXFIX to remind me to come back and make
the interpolator table static. There is no need for
a separate instance for every single Interpolator object.