Tube Unit: starting to rework hide/show logic in panel.
I discovered that the way I hide/show the VENT/SEAL toggle
and the audio emphasis path are flawed when there is
more than one Tube Unit instance in a patch.

The problem is that VCV Rack caches a single instance
of an SVG that is shared among all widgets that use it.
So when one widget mutates the visibility of an SVG,
any other widget that gets repainted will render incorrectly.

It's not always obvious because I was doing the mutation
inside the draw method, which would usually work.

So I'm reworking the hide/show stuff using separate SVG files.
It looks like SvgScrew is the ideal class for doing what I
want, because it's just an SVG widget with a framebuffer.
I can hide/show at the widget level at runtime.