Tube Unit labels keep same contrast when audio path is displayed.
I moved the control group text layers into their own SVG
file, so that I can treat them as a separate layer on top
of the audio emphasis path, which has slight opacity.
The opacity is intended to highlight the colored pentagon
backgrounds, but I don't want to make the text labels
fainter in the process.