Tube Unit: Improved font legibility at low zoom levels.
The trick of using SvgScrew for layers caused horrible looking text
at low zoom levels. I suspect it has to do with lack of
oversampling that I see present in other widgets.
I changed to using SvgPanel, which for some reason requires
me to make the dimensions exactly the same as my main panel,
even for SEAL/VENT. But that's fine because most of my overlays are transparent.