Yet another attempt to get cppcheck working in GitHub Actions.
Pass in a command-line flag that tells my `runtests`
script whether it is running in the context of GitHub Actions.
If so, do not suppress the warning about missing system includes
that happens on my system but not GitHub Actions.
If I include that suppression in GitHub Actions, cppcheck
causes ANOTHER error that it doesn't need to suppress that!