Starting work on NucleusEngine minimum energy solver.
I want the NucleusEngine to start in a local minimum energy
state, so that it is quiet by default. It should only make
noise as a result of external stimulation (or perhaps internal
feedback, later).

So I am starting a solver that will run the NucleusEngine
with high friction (low decay) until its particles converge
to approximate stillness.

Then I will save that state to a header file that can be used
to initialize NucleusEngine directly to a quiet state.

None of this has been implemented yet; this is just a stub.