Nucleus: replace speed clamping with bicubic limiter.
Nucleus has a much more interesting and less harsh sound now
because I made the speed limit apply to a bicubic limiter
instead of being an abrupt discontinuity in the system behavior.

Now I need to re-evaluate the speed limit itself, along with
the range of the output level knob.