Nucleus: fixed bug with AGC and DC REJ.
When DC REJ was enabled, the AGC was being applied to
the unfiltered signal, not the filtered signal.

I decided to move DC REJ logic into the Nucleus engine
class, so that it can manage both AGC and DC REJ and make
them interact correctly.

This will also help other implementations create
outputs that are exactly like the VCV Rack implementation.