Nucleus: decrease upper range of the SPEED knob.
Instead of allowing the SPEED to increase to +7 (128 times),
allow it to increase to +5 (32 times). Redefine the range from
-7..+5 to -6..+6, to make more sense to the user.

Add support for a fixed oversampling, so that I can run
predictably over a wide range of carefully-tested SPEED values.
Enable fixed oversampling = 1 update/sample in the VCV Rack
version of Nucleus.