Tricorder: refactored communication.
Added struct Tricorder::Communicator that centralizes
the quirks of communicating with a Tricorder module
on the right side.

Soon I will also add methods for the sender to find
out whether a Tricorder actually processed its previous message.
That will allow Nucleus to conditionally render an indicator
for which output row is being sent to Tricorder.