Nucleus: fixed flickering with drawing arrow for selected output row.
The flickering problem was caused by accidentally toggling
the flag that tells Nucleus whether Tricorder is receiving its
vector messages. Now I set a countdown to 10 when Tricoder is
known to be connected, and other decrement the counter if positive.
Only when the counter reaches zero do we give up on the idea
that Tricoder might still be there.