Tricorder: improved button fade behavior
When the mouse enters Tricorder's display area, no
longer immediately show the command buttons. The problem
is that someone sweeping the mouse rapidly across a
Tricorder causes distracting flashing of the commands.
Instead, wait a little while before fading in the buttons.

Overhauled all the fading logic based on a careful review
of the entire desired state machine.