Nucleus documentation: first draft
Completed the first draft of the main documentation
page for Nucleus. I also had to update documentation
for the output level control in Elastika and Tube Unit,
because they share changes made to the AGC for Nucleus.

Added screenshots needed to assist the documentation.

Still need to make demo videos!