Elastika: share pink warning light with Nucleus.
In Nucleus I detect non-finite output and reset the engine automatically.
When I do this, I also turn on a bright pink light on the OUTPUT knob
for 1 second, so the user has some kind of feedback that something is wrong.

Now Elastika has the same capability. It was already checking for bad
output and resetting itself, but it didn't have the same warning light.