Hiss: much more CPU-efficient when few outputs are used.
Hiss now detects which output ports are not currently being used
and refrains from generating outputs for them.
This reduces CPU overhead on my development system from 2.0% to 0.4%
when only one port is connected.
I believe most of the overhead is in std::normal_distribution
calculating the shape of the bell curve.