Tin, Tout: added LEVEL control group
Added the ability to adjust the levels on Tin and Tout.
They can also be controlled via CV + attenuverter.
It's possible to use these as part of the audio signal chain,
but the VCV VCA works just as well.

Refactored all the copied versions of getControlValue() into
a new base class SapphireModule.
Some of these were not clamping, e.g. Elastika, which is scary!