Polynucleus: panel prep for aether idea
I have an idea for a kind of aether frame dragging.
This involves two additional parameters: SPIN and VISC.
VISC will default to 0, meaning there is no frame dragging effect,
just like the existing behavior of Polynucleus.
VISC is allowed to vary anywhere in the closed range [0, +1].
As VISC increases toward +1, a rotation of space around the
z-axis becomes more pronounced. This creates a tangential force
around the z-axis on each particle in proportion to how much
that particle's angular speed around the origin differs from
the second parameter SPIN.

SPIN is allowed to vary over the closed range [-1, +1].
It gets put through some function (to be determined) that
converts the SPIN value into an angular rotation rate
expressed in radians/second.