Persist the low-sensitivity setting for each Sapphire attenuverter knob.
Renamed AutomaticLimiterModule to SapphireAutomaticLimiterModule,
so it is more obvious when I'm reading the code that the class is mine!

SapphireModule base class now saves/loads JSON to preserve the low-sensitivity
knob options across all my modules that use control groups.
All derived classes need to chain to dataToJson/dataFromJson.

The integer identifiers for all attenuverter knobs that have low sensitivity
enabled get put into a list called lowSensitivityAttenuverters. If an
identifier is in this list, it means low sensitivity is enabled;
otherwise, it is disabled:

{
  "plugin": "CosineKitty-Sapphire",
  "model": "Polynucleus",
  "version": "2.4.1",
  "params": [
    {
      "value": 6.0,
      "id": 0
    },
    // ...
    {
      "value": 5.0303997993469238,
      "id": 15
    }
  ],
  "data": {
    "lowSensitivityAttenuverters": [9],     // <=== here is the new part
    "limiterWarningLight": true,
    "agcLevel": 5.0303997993469238,
    "tricorderOutputIndex": 1
  }
}