Renamed paramId to attenId to reduce making the same mistake again.
My recent bug with using paramId where I should have used attenId
motivated me to reduce the chance of that sort of thing happening
again. So I have renamed a couple of function parameters from
paramId to attenId to make it clearer that I specifically need
the ID for an attenuverter knob, not a general parameter.