Tricorder: rotation speed must compensate for frame rate changes.
This is the first change in allowing the user to adjust the
Tricorder axis rotation rate. I now compensate for changes
in the VCV Rack frame rate because they were making the
rotation speed change.

Save/preserve the rotation speed setting, even though the
user can't change it yet. That will be my next change.