Fixed #25 - Control rotation rate of Tricorder axes.
Now Tricorder's context menu includes a slider that
allows you to change the rotation rate over the
exponential range 0.01 RPM to 100 RPM. The default
value is the geometric mean = 1 RPM.

The rotation rate is not exact. I believe I will need
to use actual time measurements to correct for the
rotation rate.