Finally resolved on a pattern for handling multiple channels.
Created a Frame class that represents 0..32 channels of data per frame.
This could be anything: a stereo audio frame, a 3D vector, a 3x3 matrix, etc.
In general, a Frame represents the input or output to a given time step.
At first I was going to return a Frame, but this is one of those cases
where the overhead would be excessive. It is better to pass things by
reference, so that only the minimum amount of work needs to be done,
on a per-case basis.