Spatula: fixed bad initialization code
The DispersionBuffer class did not actually initialize
the complex factor array to all (1, 0) like I thought.
It was a little too optimized! It thought the angle
had not changed, so it never initialized the complex
numbers, which stayed at (0, 0). This caused missing
audio from bands where I didn't set the dispersion.

In this commit, I leave the Spatula engine in its default
state without any dispersion. I verified that the output
sounds just like the input, only delayed by a fraction
of a second.