Spatula: CPU optimization - only set parameters when they have effect.
The level and dispersion parameters only have effect when the
granulizer is processing a block. Therefore, it is more efficient
for Spatula to check those two control groups and update the values
right at that moment.

Added new methods isFinalFrameBeforeBlockChange to report when
the granulizer is full, forcing it to calculate a block on
the next process call().

Only query the control groups and set parameters when that method
returns true.

This change made Spatula's CPU overhead go from 4.5% to 0.8%.