Spatula engine: added BANDWIDTH parameter.
Allow calling code to adjust the relative width of
each frequency band.

Detach updating the curve from setting the sampling rate.
Now that bandwidth also affects the curves, keep a dirty
bit and only update it when something that affects the curve
is changed.

Right before processing a block, check the dirty bit and
lazily update the curve only when needed.