Spatula panel: first working draft of flat control grid.
The new flat control grid style allows for an expanding
number of adjustable parameters. Because we have a fixed
number of bands (5) and an unknown, possibly large, number
of parameters, it makes sense for bands to be arranged
in rows and parameters to be arranged in columns.

Therefore, the five bands are now arranged vertically,
and the parameters (LEVEL, DISPERSION, BANDWIDTH)
are arraged in columns.

I can keep adjusting the width of Spatula's panel
as I add or remove parameter columns.