Project Patina: experiment with Dattorro diffusers.
Based on the works of Jon Dattorro:

https://ccrma.stanford.edu/~dattorro/EffectDesignPart1.pdf

I want to re-create his diffuser chain to see what it sounds
like on its own. This might possibly end up being a proto-reverb
module if I add a little simple feedback to it. We'll see...

Anyway, this is a reproduction of his AllpassFilter.

I'm also getting ideas based on Valley Plateau, which is
based on the Dattorro plate reverb:

https://github.com/ValleyAudio/ValleyRackFree/blob/main/src/dsp/delays/AllpassFilter.hpp