Starting an experimental module: Allpass.
Allpass is not intended as a release module.
It is an experiment to help me learn more about reverb algorithms.
However, it is possible I might merge Allpass into the main
branch later and release it as a module to the public,
if it turns out to be fun and/or useful.