Starting adaptation of Airwindows Galactic.
Inspired by a conversation on a GitHub issue:
https://github.com/baconpaul/airwin2rack/issues/111

I am adapting Chris Johnson's Airwindows Galactic
reverb as a Sapphire module. Mostly I am wanting to
start developing reverb algorithms of my own, and this
is a good exercise to reverse-engineer an algorithm
I already know and love.

This is a blank panel with no ports or controls.