Frolic, Glee: decided that the chaos mode should be reset on init after all.
I kept going back and forth about whether the chaos mode is
an internal state or an externally fed parameter. The distinction
relates to whether ChaoticOscillator::initialize() should reset
the chaos mode.

I decided that, because ChaoticOscillator is the source of truth
for the chaos mode, re-initializing it should reset the mode too.
In cases where a VCV Rack widget controls part of an engine's state,
and that widget gets reset already, there is no need for the engine
to reset it, and in fact it represents a sloppy lack of clear ownership.
But this case isn't like that, because we always get/set the mode
through the engine object. There is no widget owning the state,
just a menu system that gets/sets the engine's current state.

Also, I renamed getDefaultMode to getLegacyMode, to clarify the
following distinction:

- legacy: how the module used to behave before we added chaos modes,
  and how we want to stay when we load an older patch.

- default: how we now want the module to behave by default when newly created.

In Glee, these are different now: Legacy = 2 = Cantaloupe,
but Default = 0 = Apple, which is a lot more fun than Cantaloupe.