Sapphire Galaxy: Resolve linker conflict in Cardinal builds.
In Cardinal, all the plugins are statically linked into the executable.
This means we have to be more careful about global variable conflicts.
This happened with Sapphire Galaxy, whose model's global variable
name was "modelGalaxy".

This was the same name as the variable for Amalgamated Harmonics Galaxy.
I renamed my global variable to "modelSapphireGalaxy" to resolve this issue.