Pop: made the CHAOS knob a LOT more fun and intuitive using cubic function.
By mapping y = x^3, where x is the knob position in the range [0, 1],
we get a value y in that same range [0, 1].

But it makes the knob feel a lot more "human" in a way that is hard
to articulate. The idea came to me because I found myself most
interested in the knob positions very near zero. The cubic transform
concentrates more of the knob's angular range to the region I found
myself most attracted to, while still leaving the full behavior of
the Pop engine available.