Pop: start in the trigger-fire state immediately upon reset.
Because Pop can be operated with CHAOS=0 to act as a metronome,
it is a good idea to start immediately in the trigger-fire state
when we reset.

Before this change, resetting more frequently than the repeat rate
resulted in silence.

If this behavior bothers people, I would be happy to consider making
it an option in the context menu.