Store panel coordinates in C++ code, not SVG file.
Dan Green from 4ms reported that it is difficult for his firmware
to obtain port/control coordinates from an SVG file. So I refactored
the Python code that generates the panels to put these coordinates
in a new header file sapphire_panel.hpp instead.

My workflow can still be the same, and the Python code is still
the source of truth for where everything is located.
This change greatly simplifies things because we can get rid
of all the code for parsing and scanning through SVG files.