Eliminated fossil class layer for "reloadable modules".
I no longer support the idea of reloading an SVG file
to get updated component locations. It was a fun idea
at the time, but has not actually been very useful.

It also turned out to be unhelpful for porting Sapphire
modules to 4ms.

So no more "reloadable" widget classes!
There is no longer a reason to have a C++ class devoted
to a development tool that no longer exists.
I also renamed derived type for widgets to SapphireWidget.