Gravy: adjusted the range of the RES knob.
Remap RES to concentrate more of the knob range near the high end.
This gives the perception of more smoothly variable resonance
when increasing the RES knob.

Do not allow the internal resonance parameter `k` to get too close to zero,
because it causes ringing and, when sweeping frequency, an aliasing sound.