PGravy: glowing GAIN knob only when a connected output is being limited.
Because PGravy has 3 output ports (LP, BP, HP), and each has
an independent limiter, it makes sense to turn on the warning glow
on the GAIN knob only when a connected output is being limited.
Otherwise, there can be a glow that does not correspond to any
output currently in use. This is more confusing than useful.
So only glow the knob when limits are applied to the currently
used set of output ports.