Unit test that detects inconsistent event timing in Pop.
It turns out Sapphire Pop is not exactly following my intentions
with respect to the SPEED control. It is supposed to be V/OCT,
with SPEED=0 defined as 2 Hz.

This unit test confirms there really is a problem when SPEED
is not 0. When SPEED is 0, the interval timing is exact.