Pop: make SPEED easy to operate in V/OCT mode.
I realized that it was quite difficult and annoying to tune
Pop when CHAOS=0 (or very small). I changed its CV input so
that when you turn the attenuverter all the way clockwise to +100%,
the control voltage is added with unity gain to the SPEED value.
This makes V/OCT operation easy by turning the attenuverter all
the way clockwise. Before this change, you had to figure out that
the attenuverter should be set to some weird value like 35.7%.