Hadley Attractor is not much fun.
I added the remaining two chaos modes to the Hadley Attractor.
But playing with it just doesn't seem very fun, and doesn't seem
to add much new to Frolic, Glee. Hadley was not as space-filling
as I thought, and there isn't much fun variety in the tunings.

I'm keeping this for completeness, but I'm going to start
looking at other attractors to use for Lark instead of Hadley.