Chaops: persist memory cells in Frolic/Glee/Lark json.
As weird as this sounds, it should be possible to have multiple
Frolic/Glee/Lark instances in a patch, and a single Chaops.
You should be able to move the Chaops around and attach it to
different chaos modules without it bringing memory cells with it.
The memory cells must belong to the chaos modules because their
regions of stability are not compatible with each other.

Therefore, we store the 16 memory cells inside each chaos module.
Frolic/Glee/Lark are responsible for serializing the memory cells
to JSON and deserializing them back from JSON.