Chaops store/recall now preserves/restores exact values without transforms.
I was concerned that the Remap/Invmap pair used to store/recall
the state of a chaotic oscillator was not exact and could cause
different behavior between the original trajectory and the trajectory
after a recall operation.

So I am storing/recalling the exact internal state of the chaotic
oscillator without any transformation functions getting involved.