Project Hamburger first draft
Starting "Project Hamburger": a generic framework for running
a physical model at a different sample rate from the signal flow.
The signal rate is the sample rate that a VCV Rack module is called at.
We want to enable the physical model to run at a different rate.

This is the first draft, and it compiles, but has not been
used or tested yet. My first test will be in Elastika, but I'm
considering using it for Nucleus/Polynucleus also.