Starting Elastika mesh pre-generation project.
I can tell that about half the CPU overhead in Elastika
is in the force calculator, so I'm going to focus on optimizing it.
I will try pregenerating C++ code for the force calculator based
on the known configuration of balls and springs. This should
allow me to unroll all the loops and eliminate 'if' statements.