Elastika: springs are no longer needed as a data structure.
I only need Spring and SpringList in the code generator.

The runtime PhysicsMesh/ElastikaMesh classes do not need
to represent explicit springs, because they are implicit
in the generated code. I created a new derived type PhysicsMeshGen
that has springs, but only in the code generator.

This makes the Sapphire plugin a little smaller,
make makes it use a little less memory for each Elastika instance.