ElastikaMesh memory improvements
Reserve the exact amount of space needed for originalPositions,
currBallList, nextBallList, and forceList. This prevents
re-allocations during construction.

The forceList only needs space for mobile balls at the front.
Do not allocate space in forceList for anchors.

There is no need to pass forceList as a parameter to CalcForces,
because there is only one forceList.