Env: output envelope closely matches input amplitude
Objectively speaking, when fed a pure sinewave at a given
pitch, the output envelope voltage should match the
sinewave's amplitude.

Added a unit test to verify this, then used a correction
factor inside EnvPitchDetector to bring the somewhat weak
envelope up to full power. It wasn't quite as high as it
should be due to integration jitter.