Env: added a GAIN control group for the output envelope.
When the ENV linear output mode is selected, the new
GAIN control multiplies the envelope amplitude by
a factor anywhere from 0 to 16, using a fourth-power
curve from 0 to 2. The default value is 1.

The knob is displayed in dB:
defaulting to 0 dB,
minimum of negative infinity dB,
maximum just higher than +24 dB.