Opal engine: invert polarity of the output.
In order to allow an FM input (e.g. on VCV VCO)
to use a positive attenuverter setting, I need
Opal to send the opposite polarity feedback signal.

Specifically, Opal should send a negative voltage
when the error is positive, and vice versa.
This way a positive attenuverter setting on the
feedback path (VCO FM) attracts the system toward convergence.