Fixed bugs with incorrect output channel count in polyphonic modules.
There were a couple of places where the number of channels on an
input port did not affect the number of output channels.
Refactored to use an existing method, but updated it to handle
both cases where we want the minimum number of channels to be 0,
others where we want it to be 1.