Opal: added HI CUT control group.
HI CUT adjusts the corner frequency of two filters inside
the feedback controller: the input filter that conditions
the inbound error signal, and the output filter that
smooths out any sudden changes in the control signal.

By default, the corner frequency is 16 Hz, represented by
the knob value 0. The knob is allowed any value in the range
[-4, +4], with the default value 0. This results in a frequency
range of 1 Hz to 256 Hz.

When the HI CUT attenuverter is set to +100%, the HI CUT
CV input acts on a V/OCT scale.