Factor idea of a "toggle group" from Chaops/FREEZE.
I like how the FREEZE control works in Chaops:
it is a port and a button connected by a horizontal line.
The button toggles the state, and so does a gate fed
to the port. The two are XOR-ed to allow easy and flexible
operation.

This commit factors out the SVG generation + control locations
in generating a control like this, which I will call a
"toggle group".

There will be more commits to streamline the adding of new
toggle groups.

Then I will create a new toggle group for Opal: ACTIVE,
to enable/disable the controller from struggling to stabilize
the system.