Opal: added ENABLE toggle group.
Factored out the concept of a "toggle group" from the Chaops FREEZE control.
Added a toggle group in Opal to enable/disable the feedback controller.
When disabled, prevent the controller from "fighting" with the rest of the patch.
This allows the user to set things up to try to be tunable in the first place,
quickly enable Opal, see what happens, and maybe disable and try again.