MultiTap: create expander on button release, not button press
It is better to trigger the creation of an expander when
the insert button is released, rather than when it is pressed.

One reason is that it is slightly startling, because the action
happens before people expect.

A more important reason is that if you press and hold the button,
then change your mind, you can slide the mouse away from the
button before releasing the mouse click. Now, no expander will
be created. It is more of a definite intention when the user
hovers the mouse over the button while both clicking and releasing
the mouse.